Move all external command execs to common file
